Undo namespace rename
