+++ Emerald/App.xaml
@@ -1,4 +1,4 @@
-﻿<Application x:Class="Emerald.Uno.App"
+﻿<Application x:Class="Emerald.App"
        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
