+++ Emerald/MainPage.xaml
@@ -1,12 +1,12 @@
-﻿<Page x:Class="Emerald.Uno.MainPage"
+﻿<Page x:Class="Emerald.MainPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:local="using:Emerald.Uno"
+      xmlns:local="using:Emerald"
       xmlns:utu="using:Uno.Toolkit.UI"
-      xmlns:models="using:Emerald.Uno.Models"
-      xmlns:uc="using:Emerald.Uno.UserControls"
-    xmlns:conv="using:Emerald.Uno.Helpers.Converters"
-      xmlns:helpers="using:Emerald.Uno.Helpers">
+      xmlns:models="using:Emerald.Models"
+      xmlns:uc="using:Emerald.UserControls"
+    xmlns:conv="using:Emerald.Helpers.Converters"
+      xmlns:helpers="using:Emerald.Helpers">
   <Page.Resources>
     <conv:BoolToVisibility x:Key="BoolToVisibilty"/>
   </Page.Resources>
