+++ Emerald/Models/SquareNavigationViewItem.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel;
 using CommunityToolkit.Mvvm.ComponentModel;
 
-namespace Emerald.Uno.Models;
+namespace Emerald.Models;
 
 public partial class SquareNavigationViewItem : Model
 {
