+++ Emerald/Package.appxmanifest
@@ -7,13 +7,13 @@
   IgnorableNamespaces="uap rescap">
 
   <Identity
-    Name="Emerald.Uno"
+    Name="Emerald"
     Publisher="O=Riverside Valley"
     Version="1.0.0.0" />
 
   <Properties>
-    <DisplayName>Emerald.Uno</DisplayName>
-    <PublisherDisplayName>Emerald.Uno</PublisherDisplayName>
+    <DisplayName>Emerald</DisplayName>
+    <PublisherDisplayName>Emerald</PublisherDisplayName>
   </Properties>
 
   <Dependencies>
@@ -30,8 +30,8 @@
       Executable="$targetnametoken$.exe"
       EntryPoint="$targetentrypoint$">
       <uap:VisualElements
-        DisplayName="Emerald.Uno"
-        Description="Emerald.Uno">
+        DisplayName="Emerald"
+        Description="Emerald">
         <uap:SplashScreen />
       </uap:VisualElements>
     </Application>
