+++ Emerald/Properties/launchSettings.json
@@ -9,21 +9,21 @@
   },
   "profiles": {
     // Note: In order to select this profile, you'll need to comment the `Packaged` profile below until this is fixed: https://aka.platform.uno/wasdk-maui-debug-profile-issue
-    "Emerald.Uno (WinAppSDK Unpackaged)": {
+    "Emerald (WinAppSDK Unpackaged)": {
       "commandName": "Project",
       "compatibleTargetFramework": "windows"
     },
-    "Emerald.Uno (WinAppSDK Packaged)": {
+    "Emerald (WinAppSDK Packaged)": {
       "commandName": "MsixPackage",
       "compatibleTargetFramework": "windows"
     },
-    "Emerald.Uno (Desktop)": {
+    "Emerald (Desktop)": {
       "commandName": "Project",
       "compatibleTargetFramework": "desktop"
     },
-    "Emerald.Uno (Desktop WSL2)": {
+    "Emerald (Desktop WSL2)": {
       "commandName": "WSL2",
-      "commandLineArgs": "{ProjectDir}/bin/Debug/net8.0-desktop/Emerald.Uno.dll",
+      "commandLineArgs": "{ProjectDir}/bin/Debug/net8.0-desktop/Emerald.dll",
       "distributionName": "",
       "compatibleTargetFramework": "desktop"
     }
