+++ Emerald/Strings/en/Resources.resw
@@ -118,6 +118,6 @@
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
   <data name="ApplicationName" xml:space="preserve">
-    <value>Emerald.Uno-en</value>
+    <value>Emerald-en</value>
   </data>
 </root>
