+++ Emerald/UserControls/Titlebar.xaml
@@ -1,8 +1,8 @@
 ﻿<UserControl
-    x:Class="Emerald.Uno.UserControls.TitleBar"
+    x:Class="Emerald.UserControls.TitleBar"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Emerald.Uno.UserControls"
+    xmlns:local="using:Emerald.UserControls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
@@ -27,7 +27,7 @@
                 Margin="8,0,0,0"
                 VerticalAlignment="Center"
                 Style="{ThemeResource CaptionTextBlockStyle}"
-                Text="Emerald.Uno" />
+                Text="Emerald" />
       <TextBlock
                 x:Name="PreviewTitle"
                 Margin="4,0,0,0"
