+++ Emerald/UserControls/Titlebar.xaml.cs
@@ -15,7 +15,7 @@ using Windows.Foundation.Collections;
 
 // The User Control item template is documented at https://go.microsoft.com/fwlink/?LinkId=234236
 
-namespace Emerald.Uno.UserControls;
+namespace Emerald.UserControls;
 public sealed partial class TitleBar : UserControl
 {
     public TitleBar()
